LINARO: Disable CONFIG_OMAP2_DSS_SDI
This allows display to work on omap4/panda.
Side effect is that panel-acx565akm module is disabled
as well because it needs DSS_SDI.

This is actually not just a workaround but the proper
fix for now.  If a board has an sdi display then the
sdi requlator shoulc be enabled in sdi.c not in the
board file.

Signed-off-by: John Rigby <john.rigby@linaro.org>