LINARO: SAUCE: OMAP4: clock: wait for module to become accessible on a clk enable
On omap4, the clock f/w is missing a check for module accessebility
on a call to clk enable. Since this check is already part of the
omap device layer, this check at the clock framework is redundant.
But since not all drivers are yet using omap device apis, and most still
depend on clk framework to enable clocks, this patch seems to be needed
at this point to avoid any module accessebility issues

Signed-off-by: Rajendra Nayak <rnayak@ti.com>